basecomponents variable not found
